Implement random beatmap card
