+++ osu.Game/Screens/OnlinePlay/Matchmaking/Match/BeatmapSelect/SubScreenBeatmapSelect.cs
@@ -7,6 +7,7 @@ using osu.Framework.Extensions.ObjectExtensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Online.Multiplayer;
+using osu.Game.Online.Multiplayer.MatchTypes.Matchmaking;
 using osu.Game.Online.Rooms;
 
 namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
@@ -58,6 +59,7 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
 
             client.MatchmakingItemSelected += onItemSelected;
             client.MatchmakingItemDeselected += onItemDeselected;
+            client.SettingsChanged += onSettingsChanged;
         }
 
         private void onItemAdded(MultiplayerPlaylistItem item) => Scheduler.Add(() =>
@@ -80,6 +82,17 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
             beatmapSelectGrid.SetUserSelection(user, itemId, false);
         }
 
+        private void onSettingsChanged(MultiplayerRoomSettings settings)
+        {
+            if (client.Room!.MatchState is not MatchmakingRoomState matchmakingState)
+                return;
+
+            if (matchmakingState.CandidateItem != -1 || client.Room!.CurrentPlaylistItem.Expired)
+                return;
+
+            beatmapSelectGrid.RevealRandomItem(client.Room!.CurrentPlaylistItem);
+        }
+
         public void RollFinalBeatmap(long[] candidateItems, long finalItem) => beatmapSelectGrid.RollAndDisplayFinalBeatmap(candidateItems, finalItem);
 
         protected override void Dispose(bool isDisposing)
@@ -91,6 +104,7 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
                 client.ItemAdded -= onItemAdded;
                 client.MatchmakingItemSelected -= onItemSelected;
                 client.MatchmakingItemDeselected -= onItemDeselected;
+                client.SettingsChanged -= onSettingsChanged;
             }
         }
     }
